requirements.txt
torchvision==0.9.1+cu102
adabelief_pytorch==0.1.0
bunch==1.0.1
matplotlib==3.1.1
torch==1.8.1+cu102
scikit_image==0.15.0
kornia==0.4.1
tqdm==4.51.0
numpy==1.16.6
packaging==20.4
simplejson==3.17.0
fvcore==0.1.1
opencv_python==3.4.3.18
Pillow==8.2.0
PyYAML==5.4.1
scikit_learn==0.24.2
SimpleITK==2.0.2
skimage==0.0
